프로젝트 깃허브 업로드

+++ app/src/main/java/com/psg2024/tpprototypeapp/activities/MainActivity2.kt
@@ -33,7 +33,7 @@ class MainActivity2 : AppCompatActivity() {
                 .show()
         }
     private fun requestMyLocation() {
-        //요청 객체 생성
+
         val request: LocationRequest =
             LocationRequest.Builder(Priority.PRIORITY_HIGH_ACCURACY, 3000).build()
         //실시간 위치정보 갱신 요청- 퍼미션 코드 체크가 있어야만 함
@@ -54,7 +54,7 @@ class MainActivity2 : AppCompatActivity() {
         setContentView(binding.root)
         if(G.collectionName == null){
             G.collectionName = getSharedPreferences(this, "collectionName", "").toString()
-            if (G.collectionName != "" && G.collectionName!!.split(",")[0]==G.userAccount!!.ID ){
+            if (!G.collectionName.isNullOrEmpty() && G.userAccount != null && G.collectionName!!.split(",")[0] == G.userAccount!!.ID ){
                 AlertDialog.Builder(this).setMessage("이전에 생성한 방이 있습니다. 이어서 하시겠습니까?").setNegativeButton("취소"){_,_->
                     //firestore 정보 삭제
                     Firebase.firestore.collection(G.collectionName!!).document(G.userAccount!!.ID).delete()
@@ -70,10 +70,10 @@ class MainActivity2 : AppCompatActivity() {
         }
 
         val permissionState: Int =
-            checkSelfPermission(android.Manifest.permission.ACCESS_FINE_LOCATION)
+            checkSelfPermission(Manifest.permission.ACCESS_FINE_LOCATION)
         if (permissionState == PackageManager.PERMISSION_DENIED) {
             //퍼미션을 요청하는 다이얼로그 보이고, 그 결과를 받아오는 작업을 대신해주는 대행사 이용
-            permissionResultLauncher.launch(android.Manifest.permission.ACCESS_FINE_LOCATION)
+            permissionResultLauncher.launch(Manifest.permission.ACCESS_FINE_LOCATION)
         }
         clickBtn()
         confirmInvite()
@@ -102,7 +102,7 @@ class MainActivity2 : AppCompatActivity() {
     private fun confirmInvite() {
         val db = Firebase.firestore
         db.collection("InviteFriendlist").document(G.userAccount!!.ID).get().addOnSuccessListener {
-            if(it.exists()){
+            if(it.exists()&& G.userAccount != null){
                 //초대 받은 목록이 있으면
                 val inviteroom = it.data?.get("초대된 방") as String
                 //초대된 방 문자열에서 문자 추출
@@ -119,8 +119,12 @@ class MainActivity2 : AppCompatActivity() {
                         db.collection("InviteFriendlist").document(G.userAccount!!.ID).delete()
                         G.collectionName = inviteroom
                         db.collection(inviteroom).document(inviter).get().addOnSuccessListener {
-                            G.pos.add(it.get("LoLat").toString())
-                            G.pos.add(it.get("LoLong").toString())
+                            val loLat = it.get("LoLat") as? String
+                            val loLong = it.get("LoLong") as? String
+                            if (loLat != null && loLong != null) {
+                                G.pos.add(loLat)
+                                G.pos.add(loLong)
+                            }
                             val userInformation: MutableMap<String, Any> = mutableMapOf()
                             userInformation["ID"] = G.userAccount!!.ID
                             userInformation["Lat"] =0
@@ -142,7 +146,7 @@ class MainActivity2 : AppCompatActivity() {
     }
     // 값 불러오기
     fun getSharedPreferences(context: Context, key: String, defaultValue: Any): Any {
-        val sharedPreferences = context.getSharedPreferences("YOUR_APP_NAME", Context.MODE_PRIVATE)
+        val sharedPreferences = context.getSharedPreferences("너 어디야?", Context.MODE_PRIVATE)
         return when (defaultValue) {
             is String -> sharedPreferences.getString(key, defaultValue) ?: ""
             is Int -> sharedPreferences.getInt(key, defaultValue)
@@ -153,7 +157,7 @@ class MainActivity2 : AppCompatActivity() {
         }
     }
     fun deleteSharedPreferences(context: Context, key: String) {
-        val sharedPreferences = context.getSharedPreferences("YOUR_APP_NAME", Context.MODE_PRIVATE)
+        val sharedPreferences = context.getSharedPreferences("너 어디야?", Context.MODE_PRIVATE)
         val editor = sharedPreferences.edit()
         editor.remove(key)
         editor.apply() // 비동기적으로 삭제
